+++ src/ts/Ads/Views/OpenMeasurement/OpenMeasurement.ts
@@ -337,7 +337,7 @@ export class OpenMeasurement<T extends Campaign> extends View<T> {
              * admob-session-interface - calls session start for admob
              * vast video event handler - calls session start for vast
              */
-            if (CustomFeatures.isIASVendor(vendorKey) || this._campaign instanceof AdMobCampaign) {
+            if (CustomFeatures.isWhitelistedOMVendor(vendorKey) || this._campaign instanceof AdMobCampaign) {
                 if (this._sessionStartEventData) {
                     this.sessionStart(this._sessionStartEventData);
                 }
@@ -363,16 +363,9 @@ export class OpenMeasurement<T extends Campaign> extends View<T> {
 
                 this.impression(this.buildVastImpressionValues(MediaType.VIDEO, AccessMode.LIMITED, screenWidth, screenHeight));
 
-                if (CustomFeatures.isIASVendor(vendorKey)) {
+                if (CustomFeatures.isWhitelistedOMVendor(vendorKey)) {
                     this.sendIASEvents(IASScreenWidth, IASScreenHeight);
                 }
-
-                this.loaded({
-                    isSkippable: this._placement.allowSkip(),
-                    skipOffset: this._placement.allowSkipInSeconds(),
-                    isAutoplay: true,                   // Always autoplay for video
-                    position: VideoPosition.STANDALONE  // Always standalone video
-                });
             });
     }